fix: remove duplicate files; remove duplicate rows in sql
